// src/api/generated/model/car.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { CarStatus } from './carStatus';

export interface Car {
  id?: number;
  licensePlate?: string;
  model?: string;
  brand?: string;
  horsepower?: string;
  status?: CarStatus;
  createdAt?: string;
  updatedAt?: string;
}