// src/api/generated/model/createDailyRevenueRequestDriverRemunerationType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */

export type CreateDailyRevenueRequestDriverRemunerationType = typeof CreateDailyRevenueRequestDriverRemunerationType[keyof typeof CreateDailyRevenueRequestDriverRemunerationType];


export const CreateDailyRevenueRequestDriverRemunerationType = {
  PERCENTAGE_SHARE: 'PERCENTAGE_SHARE',
  WEEKLY_FIXED_RATE: 'WEEKLY_FIXED_RATE',
  FLAT_RATE: 'FLAT_RATE',
} as const;